Add possibility to ignore notifications from post-commit hooks
This can be useful if you want to prevent some long-running jobs (e.g.
reports) starting because of every commit, but still want to run them
periodic if SCM changes have occurred.

By default all scm polling jobs will trigger on commit notification
(this was previous default behaviour and it was not changed).